+++ src/test/java/com/dd/plist/test/NSNumberTest.java
@@ -13,67 +13,66 @@ import static org.hamcrest.Matchers.containsString;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
- * Tests for the NSString class.
- *
+ * Tests for the {@link NSNumber} class.
  * @author Daniel Dreibrodt
  */
 public class NSNumberTest {
     @Test
-    public void testInitializeFromNanString() {
+    public void init_canHandleNaNString() {
         NSNumber nan = new NSNumber("nan");
         assertTrue(Double.isNaN(nan.doubleValue()));
     }
 
     @Test
-    public void testInitializeFromDoubleNaN() {
+    public void init_canHandleNaNDouble() {
         NSNumber nan = new NSNumber(Double.NaN);
         assertTrue(Double.isNaN(nan.doubleValue()));
     }
 
     @Test
-    public void testInitializeFromFloatNaN() {
+    public void init_canHandleNaNFloat() {
         NSNumber nan = new NSNumber(Float.NaN);
         assertTrue(Double.isNaN(nan.doubleValue()));
     }
 
     @Test
-    public void testNanIntValue() {
+    public void intValue_throwsIllegalStateExceptionForNaN() {
         NSNumber nan = new NSNumber(Double.NaN);
         assertThrows(IllegalStateException.class, nan::intValue);
     }
 
     @Test
-    public void testNanLongValue() {
+    public void longValue_throwsIllegalStateExceptionForNaN() {
         NSNumber nan = new NSNumber(Double.NaN);
         assertThrows(IllegalStateException.class, nan::longValue);
     }
 
     @Test
-    public void testNanFloatValue() {
+    public void floatValue_canHandleNaN() {
         NSNumber nan = new NSNumber(Double.NaN);
         assertTrue(Float.isNaN(nan.floatValue()));
     }
 
     @Test
-    public void testNanStringValue() {
+    public void stringValue_canHandleNaN() {
         NSNumber nan = new NSNumber(Double.NaN);
         assertEquals("nan", nan.stringValue());
     }
 
     @Test
-    public void testNanToXml() {
+    public void toXML_canHandleNaN() {
         NSNumber nan = new NSNumber(Double.NaN);
         assertThat(nan.toXMLPropertyList(), containsString("<real>nan</real>"));
     }
 
     @Test
-    public void testNanToString() {
+    public void toString_canHandleNaN() {
         NSNumber nan = new NSNumber(Double.NaN);
         assertEquals("nan", nan.toString());
     }
 
     @Test
-    public void testNanToBinaryAndBack() throws Exception {
+    public void toBinary_canHandleNaN() throws Exception {
         NSDictionary dict = new NSDictionary();
         dict.put("NaN", new NSNumber(Double.NaN));
 
@@ -91,61 +90,61 @@ public class NSNumberTest {
     }
 
     @Test
-    public void testInitializeFromPositiveInfinityString() {
+    public void init_canHandlePositiveInfinityString() {
         NSNumber inf = new NSNumber("+infinity");
         assertEquals(Double.POSITIVE_INFINITY, inf.doubleValue());
     }
 
     @Test
-    public void testInitializeFromDoublePositiveInfinity() {
+    public void init_canHandlePositiveInfinityDouble() {
         NSNumber inf = new NSNumber(Double.POSITIVE_INFINITY);
         assertEquals(Double.POSITIVE_INFINITY, inf.doubleValue());
     }
 
     @Test
-    public void testInitializeFromFloatPositiveInfinity() {
+    public void init_canHandlePositiveInfinityFloat() {
         NSNumber inf = new NSNumber(Float.POSITIVE_INFINITY);
         assertEquals(Double.POSITIVE_INFINITY, inf.doubleValue());
     }
 
     @Test
-    public void testPositiveInfinityIntValue() {
+    public void intValue_throwsIllegalStateExceptionForPositiveInfinity() {
         NSNumber inf = new NSNumber(Double.POSITIVE_INFINITY);
         assertThrows(IllegalStateException.class, inf::intValue);
     }
 
     @Test
-    public void testPositiveInfinityLongValue() {
+    public void longValue_throwsIllegalStateExceptionForPositiveInfinity() {
         NSNumber inf = new NSNumber(Double.POSITIVE_INFINITY);
         assertThrows(IllegalStateException.class, inf::longValue);
     }
 
     @Test
-    public void testPositiveInfinityFloatValue() {
+    public void floatValue_canHandlePositiveInfinity() {
         NSNumber inf = new NSNumber(Double.POSITIVE_INFINITY);
         assertEquals(Float.POSITIVE_INFINITY, inf.floatValue());
     }
 
     @Test
-    public void testPositiveInfinityStringValue() {
+    public void stringValue_canHandlePositiveInfinity() {
         NSNumber inf = new NSNumber(Double.POSITIVE_INFINITY);
         assertEquals("+infinity", inf.stringValue());
     }
 
     @Test
-    public void testPositiveInfinityToXml() {
+    public void toXml_canHandlePositiveInfinity() {
         NSNumber inf = new NSNumber(Double.POSITIVE_INFINITY);
         assertThat(inf.toXMLPropertyList(), containsString("<real>+infinity</real>"));
     }
 
     @Test
-    public void testPositiveInfinityToString() {
+    public void toString_canHandlePositiveInfinity() {
         NSNumber inf = new NSNumber(Double.POSITIVE_INFINITY);
         assertEquals("+infinity", inf.toString());
     }
 
     @Test
-    public void testPositiveInfinityToBinaryAndBack() throws Exception {
+    public void toBinary_canHandlePositiveInfinity() throws Exception {
         NSDictionary dict = new NSDictionary();
         dict.put("inf", new NSNumber(Double.POSITIVE_INFINITY));
 
@@ -163,61 +162,61 @@ public class NSNumberTest {
     }
 
     @Test
-    public void testInitializeFromNegativeInfinityString() {
+    public void init_canHandleNegativeInfinityString() {
         NSNumber inf = new NSNumber("-infinity");
         assertEquals(Double.NEGATIVE_INFINITY, inf.doubleValue());
     }
 
     @Test
-    public void testInitializeFromDoubleNegativeInfinity() {
+    public void init_canHandleNegativeInfinityDouble() {
         NSNumber inf = new NSNumber(Double.NEGATIVE_INFINITY);
         assertEquals(Double.NEGATIVE_INFINITY, inf.doubleValue());
     }
 
     @Test
-    public void testInitializeFromFloatNegativeInfinity() {
+    public void init_canHandleNegativeInfinityFloat() {
         NSNumber inf = new NSNumber(Float.NEGATIVE_INFINITY);
         assertEquals(Double.NEGATIVE_INFINITY, inf.doubleValue());
     }
 
     @Test
-    public void testNegativeInfinityIntValue() {
+    public void intValue_throwsIllegalStateExceptionForNegativeInfinity() {
         NSNumber inf = new NSNumber(Double.NEGATIVE_INFINITY);
         assertThrows(IllegalStateException.class, inf::intValue);
     }
 
     @Test
-    public void testNegativeInfinityLongValue() {
+    public void longValue_throwsIllegalStateExceptionForNegativeInfinity() {
         NSNumber inf = new NSNumber(Double.NEGATIVE_INFINITY);
         assertThrows(IllegalStateException.class, inf::longValue);
     }
 
     @Test
-    public void testNegativeInfinityFloatValue() {
+    public void floatValue_canHandleNegativeInfinity() {
         NSNumber inf = new NSNumber(Double.NEGATIVE_INFINITY);
         assertEquals(Float.NEGATIVE_INFINITY, inf.floatValue());
     }
 
     @Test
-    public void testNegativeInfinityStringValue() {
+    public void stringValue_canHandleNegativeInfinity() {
         NSNumber inf = new NSNumber(Double.NEGATIVE_INFINITY);
         assertEquals("-infinity", inf.stringValue());
     }
 
     @Test
-    public void testNegativeInfinityToXml() {
+    public void toXml_canHandleNegativeInfinity() {
         NSNumber inf = new NSNumber(Double.NEGATIVE_INFINITY);
         assertThat(inf.toXMLPropertyList(), containsString("<real>-infinity</real>"));
     }
 
     @Test
-    public void testNegativeInfinityToString() {
+    public void toString_canHandleNegativeInfinity() {
         NSNumber inf = new NSNumber(Double.NEGATIVE_INFINITY);
         assertEquals("-infinity", inf.toString());
     }
 
     @Test
-    public void testNegativeInfinityToBinaryAndBack() throws Exception {
+    public void toBinary_canHandleNegativeInfinity() throws Exception {
         NSDictionary dict = new NSDictionary();
         dict.put("inf", new NSNumber(Double.NEGATIVE_INFINITY));
